Add input checking, set defaults in config file, window resizing, header strings
Add rudimentary checking to the values stored in the config file.

Inserted sane defaults for the config file.  This fixes a bug where if you forget to set the port the program won't start.

Added automatic resizing to adjust for the terminal size.

Moved the version info to a single spot at the top of the file so multiple locations don't need to be updated for every version change.